Add super simple structured queries using JSON

+++ src/types/timestamp.rs
@@ -1,8 +1,8 @@
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq, Hash, PartialOrd, Ord, Copy)]
-pub struct Timestamp(pub u128);
+pub struct Timestamp(pub u64);
 
 impl Timestamp {
-    pub fn as_nanos(&self) -> u128 {
+    pub fn as_nanos(&self) -> u64 {
         self.0
     }
 }
